+++ SRC/cheequb.f
@@ -263,7 +263,7 @@
 *        avg = s^T beta / n
          AVG = 0.0E0
          DO I = 1, N
-            AVG = AVG + S( I )*WORK( I )
+            AVG = AVG + REAL( S( I )*WORK( I ) )
          END DO
          AVG = AVG / N
 
@@ -280,8 +280,8 @@
             T = CABS1( A( I, I ) )
             SI = S( I )
             C2 = ( N-1 ) * T
-            C1 = ( N-2 ) * ( WORK( I ) - T*SI )
-            C0 = -(T*SI)*SI + 2*WORK( I )*SI - N*AVG
+            C1 = REAL( ( N-2 ) * ( WORK( I ) - T*SI ) )
+            C0 = REAL( -(T*SI)*SI + 2*WORK( I )*SI - N*AVG )
             D = C1*C1 - 4*C0*C2
 
             IF ( D .LE. 0 ) THEN
@@ -316,7 +316,7 @@
                END DO
             END IF
 
-            AVG = AVG + ( U + WORK( I ) ) * D / N
+            AVG = AVG + REAL( ( U + WORK( I ) ) * D / N )
             S( I ) = SI
          END DO
       END DO
